disable auto detect porter.yaml on dashboard and organization
